requirements.txt
mcp>=1.2.0
netmiko
pydantic
python-dotenv
